chore: moved config to uv (#30)
* chore: moved config to uv

* doc: fixed doc tests

* fix: fixed modern numpy bug

* chore: fail workflows on error

* fix: fixed numpy related bug

* fix: fixed slow bug

* test: fixed snipet test

* doc: fixed README

* chore: remove azure devops config

* chore: fixed running examples

+++ src/pyquasoare/models.py
@@ -121,7 +121,7 @@ def quad_model(alphas, flux_scalings, coefs_noscaling,
            [ 0.00000000e+00, -1.89470944e-02]])
     >>> # Check mass balance
     >>> fx.sum(axis=0)[1]+s0-s1[-1]
-    -100.0
+    np.float64(-100.0)
     """
     if errors not in ERRORS:
         txt = "/".join(ERRORS)
